Remove lib
Generic graph based solution (at least the one I implemented) would not
be able to fit into flash and it is making it difficult to split
ownership of input/output/audio.

The purpose of this library should be to accelerate prototyping of
modules that are to be standalone and hardcoded.

Signed-off-by: Petr Horacek <hrck@protonmail.com>